move some event handlers to chrome-events
