+++ shared/js/background/companies.es6.js
@@ -145,13 +145,6 @@ var Companies = (() => {
 
 Companies.buildFromStorage()
 
-// sync data to storage when a tab finishes loading
-chrome.tabs.onUpdated.addListener( (id,info) => {
-    if (info.status === "complete") {
-        Companies.syncToStorage()
-    }
-})
-
 chrome.runtime.onMessage.addListener((req, sender, res) => {
     if (req.getTopBlocked) {
         res(Companies.getTopBlocked(req.getTopBlocked))
